Reduced main file size
