+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "python.pythonPath": "/bin/python3",
+    //"workbench.preferredLightColorTheme": "Blakai Theme",
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true
+}
